Fri, Jan 27, 2023, 5:45 PM +02:00

+++ readability/readability.c
@@ -27,8 +27,8 @@ int index;
 float L, S;
 
 
-    L = (letters / words) * 100;
-    S = (sentences / words) * 100;
+    L = (letters / (float)words) * 100;
+    S = (sentences / (float)words) * 100;
     index = (0.0588 * L ) - (0.296 * S ) - 15.8;
     printf("L:%f",L);
     printf("S:%f",S);
